+++ src/app/routes/broker-ui/catalog-page/catalog-page/catalog-page.component.ts
@@ -1,11 +1,12 @@
 import {Component, HostBinding, OnDestroy, OnInit} from '@angular/core';
 import {FormControl} from '@angular/forms';
 import {PageEvent} from '@angular/material/paginator';
-import {ActivatedRoute, Router} from '@angular/router';
+import {ActivatedRoute, Params, Router} from '@angular/router';
 import {BehaviorSubject, Subject} from 'rxjs';
-import {filter, map, take, takeUntil} from 'rxjs/operators';
+import {filter, map, takeUntil} from 'rxjs/operators';
 import {Store} from '@ngxs/store';
 import {CatalogPageSortingItem} from '@sovity.de/broker-server-client';
+import {unique} from 'src/app/core/utils/array-utils';
 import {LocalStoredValue} from 'src/app/core/utils/local-stored-value';
 import {AssetDetailDialogDataService} from '../../../../component-library/catalog/asset-detail-dialog/asset-detail-dialog-data.service';
 import {AssetDetailDialogService} from '../../../../component-library/catalog/asset-detail-dialog/asset-detail-dialog.service';
@@ -15,6 +16,7 @@ import {
 } from '../../../../component-library/catalog/view-selection/view-mode-enum';
 import {BrokerServerApiService} from '../../../../core/services/api/broker-server-api.service';
 import {FilterBoxItem} from '../filter-box/filter-box-item';
+import {FilterBoxModel} from '../filter-box/filter-box-model';
 import {FilterBoxVisibleState} from '../filter-box/filter-box-visible-state';
 import {CatalogActiveFilterPill} from '../state/catalog-active-filter-pill';
 import {CatalogPage} from '../state/catalog-page-actions';
@@ -41,7 +43,6 @@ export class CatalogPageComponent implements OnInit, OnDestroy {
     'brokerui.viewMode',
     isViewMode,
   );
-  initialConnectorEndpointsFilterValues: string[] | null = null;
   private fetch$ = new BehaviorSubject(null);
 
   // only tracked to prevent the component from resetting
@@ -58,10 +59,10 @@ export class CatalogPageComponent implements OnInit, OnDestroy {
 
   ngOnInit(): void {
     this.store.dispatch(CatalogPage.Reset);
+    this.setConnectorEndpointFiltersFromQueryParamsOnce();
     this.startListeningToStore();
     this.startEmittingSearchText();
     this.startEmittingSortBy();
-    this.parseConnectorEndpointQueryParams();
   }
 
   private startListeningToStore() {
@@ -78,13 +79,9 @@ export class CatalogPageComponent implements OnInit, OnDestroy {
           this.sortBy.setValue(state.activeSorting);
         }
 
-        if (this.state.isPageReady) {
-          this.setInitialConnectorEndpointFilters();
-
-          if (!this.expandedFilterId) {
-            this.expandedFilterId =
-              this.state.fetchedData.data.availableFilters.fields[0].id;
-          }
+        if (this.state.isPageReady && !this.expandedFilterId) {
+          this.expandedFilterId =
+            this.state.fetchedData.data.availableFilters.fields[0].id;
         }
       });
   }
@@ -109,37 +106,47 @@ export class CatalogPageComponent implements OnInit, OnDestroy {
       });
   }
 
-  private parseConnectorEndpointQueryParams() {
-    this.route.queryParams.pipe(take(1)).subscribe((x) => {
-      if (!('connectorEndpoint' in x)) {
-        return;
-      }
-      const endpoints = x.connectorEndpoint;
-      this.initialConnectorEndpointsFilterValues = Array.isArray(endpoints)
-        ? endpoints.filter((item, i, arr) => arr.indexOf(item) === i)
-        : [endpoints];
-      // remove query params from url
-      this.router.navigate([]);
-    });
+  private setConnectorEndpointFiltersFromQueryParamsOnce() {
+    const connectorEndpoints = this.parseConnectorEndpoints(
+      this.route.snapshot.queryParams,
+    );
+    if (!connectorEndpoints?.length) {
+      return;
+    }
+
+    this.store.dispatch(
+      new CatalogPage.AddFilterBox(
+        this.buildConnectorEndpointFilterBoxModel(connectorEndpoints),
+      ),
+    );
+    this.expandedFilterId = 'connectorEndpoint';
+    // remove query params from url
+    this.router.navigate([]);
   }
 
-  private setInitialConnectorEndpointFilters() {
-    if (this.initialConnectorEndpointsFilterValues?.length) {
-      const filterBoxItems = this.initialConnectorEndpointsFilterValues.map(
-        (x): FilterBoxItem => ({type: 'ITEM', id: x, label: x}),
-      );
-      this.initialConnectorEndpointsFilterValues = null;
-      this.store.dispatch(
-        new CatalogPage.UpdateFilterSelectedItems(
-          'connectorEndpoint',
-          filterBoxItems,
-        ),
-      );
-      this.expandedFilterId =
-        this.state.fetchedData.data.availableFilters.fields.find(
-          (x) => x.id === 'connectorEndpoint',
-        )!.id;
+  private parseConnectorEndpoints(params: Params): string[] {
+    if (!('connectorEndpoint' in params)) {
+      return [];
     }
+    const endpoints = params.connectorEndpoint;
+    return Array.isArray(endpoints) ? endpoints.filter(unique) : [endpoints];
+  }
+
+  private buildConnectorEndpointFilterBoxModel(
+    endpoints: string[],
+  ): FilterBoxModel {
+    const items: FilterBoxItem[] = endpoints.map((x) => ({
+      type: 'ITEM',
+      id: x,
+      label: x,
+    }));
+    return {
+      id: 'connectorEndpoint',
+      title: 'Connector',
+      selectedItems: items,
+      availableItems: items,
+      searchText: '',
+    };
   }
 
   onDataOfferClick(dataOffer: CatalogDataOfferMapped) {